chore: remove barrel file from types because of convention violation
